Refactoring and removing the support for the sched_switch kernel tracepoint in favour of a generalization for all tracepoints
