+++ OrbitLinuxTracing/PerfEvent.cpp
@@ -42,6 +42,4 @@ void AmdgpuSchedRunJobPerfEvent::Accept(PerfEventVisitor* visitor) { visitor->vi
 
 void DmaFenceSignaledPerfEvent::Accept(PerfEventVisitor* visitor) { visitor->visit(this); }
 
-void SchedSwitchPerfEvent::Accept(PerfEventVisitor*) {}
-
 }  // namespace LinuxTracing
