+++ OrbitLinuxTracing/TracerThread.h
@@ -142,7 +142,6 @@ class TracerThread {
   absl::flat_hash_set<uint64_t> uretprobes_ids_;
   absl::flat_hash_set<uint64_t> stack_sampling_ids_;
   absl::flat_hash_set<uint64_t> task_newtask_ids_;
-  absl::flat_hash_set<uint64_t> sched_switch_ids_;
   absl::flat_hash_set<uint64_t> task_rename_ids_;
   absl::flat_hash_set<uint64_t> amdgpu_cs_ioctl_ids_;
   absl::flat_hash_set<uint64_t> amdgpu_sched_run_job_ids_;
